Paddle/python_answers/requirements.txt
pandas
requests
python-dotenv==1.0.0